# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API",
        "description": "Documentation de l'API avec Swagger",
        "contact": {
            "email": "contact@exemple.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/motorcycle-models": {
            "get": {
                "tags": [
                    "Motorcycle Models"
                ],
                "summary": "Get all motorcycle models",
                "operationId": "9cf26095369c0a1f24ccdc59eb5e8e8f",
                "responses": {
                    "200": {
                        "description": "List of motorcycle models"
                    }
                }
            },
            "post": {
                "tags": [
                    "Motorcycle Models"
                ],
                "summary": "Create a new motorcycle model",
                "operationId": "348e2a1f71de4990c263a894b77438d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "brand_id",
                                    "name",
                                    "type_id"
                                ],
                                "properties": {
                                    "brand_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "CBR600RR"
                                    },
                                    "type_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Motorcycle model created"
                    }
                }
            }
        },
        "/api/motorcycle-models/{id}": {
            "get": {
                "tags": [
                    "Motorcycle Models"
                ],
                "summary": "Get a motorcycle model by ID",
                "operationId": "46db16be3a19f6d3906d8c279c984480",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Motorcycle model data"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Motorcycle Models"
                ],
                "summary": "Update a motorcycle model",
                "operationId": "b8d3d0d6c6d3dafb99f779cc7ee53fd7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "brand_id",
                                    "name",
                                    "type_id"
                                ],
                                "properties": {
                                    "brand_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "MT-07"
                                    },
                                    "type_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Motorcycle model updated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Motorcycle Models"
                ],
                "summary": "Delete a motorcycle model",
                "operationId": "66a3f0979775886a398e833255b2b2e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/motorcycle-types": {
            "get": {
                "tags": [
                    "Motorcycle Types"
                ],
                "summary": "Get all motorcycle types",
                "operationId": "297a51041d61f5f3f541a47f0193a7cc",
                "responses": {
                    "200": {
                        "description": "List of motorcycle types"
                    }
                }
            },
            "post": {
                "tags": [
                    "Motorcycle Types"
                ],
                "summary": "Create a new motorcycle type",
                "operationId": "7ad0d99989baeb77ed42f540f0f9180d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sport"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Motorcycle type created"
                    }
                }
            }
        },
        "/api/motorcycle-types/{id}": {
            "get": {
                "tags": [
                    "Motorcycle Types"
                ],
                "summary": "Get a motorcycle type by ID",
                "operationId": "b0d8d440778b00f40145cac0d48582cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Motorcycle type data"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Motorcycle Types"
                ],
                "summary": "Update a motorcycle type",
                "operationId": "5271c1e7521208c1d4922ca3050335d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Cruiser"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Motorcycle type updated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Motorcycle Types"
                ],
                "summary": "Delete a motorcycle type",
                "operationId": "f7228e3cb76ec733860fd962dd073913",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
...
}